Move long/short channel processing from utils to channels
