+++ examples/plot_30_frequency.py
@@ -67,7 +67,7 @@ raw_intensity.annotations.delete(
 
 raw_od = mne.preprocessing.nirs.optical_density(raw_intensity)
 raw_haemo = mne.preprocessing.nirs.beer_lambert_law(raw_od)
-raw_haemo = mne_nirs.utils.get_long_channels(raw_haemo)
+raw_haemo = mne_nirs.channels.get_long_channels(raw_haemo)
 
 
 ###############################################################################
