+++ mne_nirs/__init__.py
@@ -1,5 +1,6 @@
 from ._version import __version__
 
+from . import channels
 from . import signal_enhancement
 from . import visualisation
 from . import experimental_design
