+++ mne_nirs/channels/__init__.py
@@ -1,0 +1,7 @@
+"""Functions to help with handling channel information."""
+
+# Authors: Robert Luke <mail@robertluke.net>
+#
+# License: BSD (3-clause)
+
+from ._short import get_short_channels, get_long_channels
