+++ mne_nirs/channels/_short.py
