+++ mne_nirs/utils/__init__.py
@@ -4,5 +4,4 @@
 #
 # License: BSD (3-clause)
 
-from ._short import get_short_channels, get_long_channels
 from . import _io
